Move profile and catalog pages to the separate folders
