+++ src/App.js
@@ -15,7 +15,7 @@ import useAuth from "./hooks/useAuth"
 import * as ROUTES from "./constants/routes"
 
 // import pages
-import Catalog from "./pages/catalog"
+import Catalog from "./pages/catalog/catalog"
 import Home from "./pages/home"
 import Search from "./pages/search/search"
 import Login from "./pages/login"
@@ -23,7 +23,7 @@ import SignUp from "./pages/signup"
 import NotFound from "./pages/not-found"
 import Upload from "./pages/upload"
 import Album from "./pages/album"
-import Profile from "./pages/profile"
+import Profile from "./pages/profile/profile"
 import Collection from "./pages/collection"
 import Wishlist from "./pages/wishlist"
 import MyUploads from "./pages/my-uploads"
