+++ src/pages/profile/profile.js
@@ -1,0 +1,17 @@
+import Header from "../../components/Header/Header"
+import Footer from "../../components/Footer"
+import ProfileContent from "./ProfileContent"
+
+function Profile() {
+  return (
+    <>
+      <Header />
+      <main>
+        <ProfileContent />
+      </main>
+      <Footer />
+    </>
+  )
+}
+
+export default Profile
